ДЗ 2.8 готово

+++ src/main/java/com/example/homework25/homework_2_5/data/Employee.java
@@ -26,7 +26,7 @@ public class Employee {
         return department;
     }
 
-    public double getWorkersSalary() {
+    public int getWorkersSalary() {
         return workersSalary;
     }
